+++ components/Home/RecentPosts.tsx
@@ -1,8 +1,6 @@
-import { sortPosts } from "@/lib";
-import Link from "next/link";
-import React from "react";
-import BlogLayoutThree from "../Blog/BlogLayoutThree";
 import { Post, Tag as PrismaTag } from "@prisma/client";
+import Link from "next/link";
+import BlogLayoutThree from "../Blog/BlogLayoutThree";
 
 interface PostWithTags extends Post {
   tags: PrismaTag[];
@@ -27,7 +25,7 @@ const RecentPosts = ({ posts }: RecentPostsProps) => {
     <section className="w-full mt-16 sm:mt-24 md:mt-32 px-5 sm:px-10 md:px-24 sxl:px-32 flex flex-col items-center justify-center">
       <div className="w-full flex justify-between">
         <h2 className="w-fit inline-block font-bold capitalize text-2xl md:text-4xl text-dark dark:text-light">Recent Posts</h2>
-        <Link href="/tags/all" className="inline-block font-medium text-accent dark:text-accentDark underline underline-offset-2 text-base md:text-lg">
+        <Link href="/explore" className="inline-block font-medium text-accent dark:text-accentDark underline underline-offset-2 text-base md:text-lg">
           view all
         </Link>
       </div>